Fix: refactor code shopping
