+++ resources/views/frontend/pages/shopping-cart.blade.php
@@ -1,0 +1,10 @@
+@extends('frontend.index')
+@section('content')
+
+@endsection
+
+@push('scripts')
+    <script>
+
+    </script>
+@endpush
