+++ resources/views/frontend/pages/shopping.blade.php
@@ -205,35 +205,11 @@
             }
         }
 
-        function updateCartCount() {
-            const products = JSON.parse(localStorage.getItem("products")) || {};
-            const totalCount = Object.values(products).reduce((acc, curr) => acc + curr.amount, 0);
-            const cartCount = document.getElementById("cart-count");
-            if (cartCount) {
-                cartCount.textContent = totalCount > 0 ? totalCount.toString()+'+' : "0";
-            }
-        }
-
         function updateProductAmount(productId) {
             const productAmountEl = document.querySelector(`#amount-${productId}`);
             const products = JSON.parse(localStorage.getItem("products")) || {};
             const product = products[productId];
             productAmountEl.textContent = product ? product.amount : 0;
-        }
-
-        async function productCalculation() {
-            try {
-                const response = await axios.post('{{ route('products.calculation') }}', {
-                    'products': JSON.parse(localStorage.getItem("products"))
-                });
-                const total = response.data.total
-                const cartTotalPrice = document.getElementById("cart-total-price");
-                if (cartTotalPrice) {
-                    cartTotalPrice.textContent = total > 0 ? total.toString() : "0.00";
-                }
-            } catch (error) {
-                console.error(error);
-            }
         }
 
         function updateRemoveButtonVisibility(productId) {
@@ -255,8 +231,6 @@
                     button.style.display = 'none';
                 }
             });
-            productCalculation();
-            updateCartCount()
         });
 
         const addProductButtons = document.querySelectorAll(".btn-add-product");
@@ -275,8 +249,8 @@
             localStorage.setItem("products", JSON.stringify(products));
             updateProductAmount(productId);
             updateCartCount()
-            updateRemoveButtonVisibility(productId);
             productCalculation();
+            updateRemoveButtonVisibility(productId);
         }
 
         function removeProduct(productId) {
@@ -289,8 +263,8 @@
                 localStorage.setItem("products", JSON.stringify(products));
                 updateProductAmount(productId);
                 updateCartCount();
-                updateRemoveButtonVisibility(productId);
                 productCalculation();
+                updateRemoveButtonVisibility(productId);
             }
         }
     </script>
